% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Carseats}
\alias{Carseats}
\title{Car seats dataset}
\format{A data frame with 400 rows and 11 variables:}
\source{
\url{https://rdrr.io/cran/ISLR/man/Carseats.html}
}
\usage{
Carseats
}
\description{
A simulated data set containing sales of child car seats at 400 different stores.
}
\keyword{datasets}
