% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{College}
\alias{College}
\title{College dataset}
\format{A data frame with 777 rows and 18 variables:}
\source{
\url{https://rdrr.io/cran/ISLR/man/College.html}
}
\usage{
College
}
\description{
Statistics for a large number of US Colleges from the 1995 issue of US News and World Report.
}
\keyword{datasets}
