% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Credit}
\alias{Credit}
\title{Credit dataset}
\format{A data frame with 400 rows and 10 variables:}
\source{
\url{https://rdrr.io/cran/ISLR/man/Credit.html}
}
\usage{
Credit
}
\description{
A simulated data set containing information on ten thousand customers. The aim here is to predict
which customers will default on their credit card debt.
}
\keyword{datasets}
