% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hitters}
\alias{Hitters}
\title{Hitters dataset}
\format{A data frame with 263 rows and 20 variables:}
\source{
\url{https://rdrr.io/cran/ISLR/man/Hitters.html}
}
\usage{
Hitters
}
\description{
Major League Baseball Data from the 1986 and 1987 seasons.
}
\keyword{datasets}
