% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OJ}
\alias{OJ}
\title{Credit dataset}
\format{A data frame with 1070 rows and 18 variables:}
\source{
\url{https://rdrr.io/cran/ISLR/man/OJ.html}
}
\usage{
OJ
}
\description{
The data contains 1070 purchases where the customer either 
purchased Citrus Hill or Minute Maid Orange Juice. 
A number of characteristics of the customer and product are recorded.
}
\keyword{datasets}
