\name{varselect-package}
\alias{varselect-package}
\alias{varselect}
\docType{package}
\title{
\packageTitle{varselect}
}
\description{
\packageDescription{varselect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{varselect}
\packageIndices{varselect}

}
\author{
\packageAuthor{varselect}

Maintainer: \packageMaintainer{varselect}
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
Optional links to other man pages
}
\examples{
examples
}
