% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{varselect}
\alias{varselect}
\alias{varselect.formula}
\alias{varselect.default}
\title{Varselect}
\usage{
varselect(x, ...)

\method{varselect}{formula}(fml, data, nvmax = 8)

\method{varselect}{default}(x, y, nvmax = 8)
}
\arguments{
\item{x}{optional Design matrix, coappears with \code{y}}

\item{fml}{optional formula, coappears with \code{data}}

\item{data}{Optional data frame. coappears with \code{fml}}

\item{nvmax}{Maximal number of parameters included in the subset model}

\item{y}{Response vector, coappear with \code{x}}
}
\value{
A list with multiple fields. Users can apply \code{summary}, \code{coef} and \code{plot} to explore the results.
}
\description{
Forward selection for linear regression
}
\examples{
data(Hitters)
mymodel = varselect(Salary~., data=Hitters, nvmax=12)
summary(mymodel)
coef(mymodel, id=6)
plot(mymodel)
}
