\name{varselect-package}
\alias{varselect-package}
\alias{varselect}
\docType{package}
\title{
\packageTitle{varselect}
}
\description{
\packageDescription{varselect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{varselect}
\packageIndices{varselect}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{varselect}

Maintainer: \packageMaintainer{varselect}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
